Fix return prev callback

+++ src/libtf.cpp
@@ -1,4 +1,5 @@
 #include <algorithm>
+#include <map>
 #include <numeric>
 
 #include "InputContext.hpp"
@@ -8,6 +9,21 @@
 typedef struct libtf_InputContext
 {
     CComPtr<libtf::CInputContext> m_InputContext;
+    // Candidate List
+    libtf_CandidateListCallback m_PrevCandidateListCallback = NULL;
+    void*                       m_PrevCandidateListUserData = NULL;
+    // PreEdit
+    libtf_PreEditCallback m_PrevPreEditCallback = NULL;
+    void*                 m_PrevPreEditUserData = NULL;
+    // PreEdit Rect
+    libtf_PreEditRectCallback m_PrevPreEditRectCallback = NULL;
+    void*                     m_PrevPreEditRectUserData = NULL;
+    // Commit
+    libtf_CommitCallback m_PrevCommitCallback = NULL;
+    void*                m_PrevCommitUserData = NULL;
+    // InputProcessor
+    libtf_InputProcessorCallback m_PrevInputProcessorCallback = NULL;
+    void*                        m_PrevInputProcessorUserData = NULL;
 } libtf_InputContext_t;
 /**
  * @brief Create input context for specific window
@@ -18,6 +34,8 @@ typedef struct libtf_InputContext
 LIBTF_EXPORT HRESULT libtf_create_ctx(libtf_pInputContext* ctx, const HWND hWnd)
 {
     BEGIN_HRESULT();
+
+    if (!ctx) return E_INVALIDARG;
 
     auto inputCtx = new libtf_InputContext_t();
     *ctx          = inputCtx;
@@ -64,6 +82,8 @@ LIBTF_EXPORT HRESULT libtf_set_activated(libtf_pInputContext ctx, const bool act
 {
     BEGIN_HRESULT();
 
+    if (!ctx) return E_INVALIDARG;
+
     CHECK_HR(ctx->m_InputContext->setActivated(activated));
 
     END_HRESULT();
@@ -77,6 +97,8 @@ LIBTF_EXPORT HRESULT libtf_set_activated(libtf_pInputContext ctx, const bool act
 LIBTF_EXPORT HRESULT libtf_get_activated(const libtf_pInputContext ctx, bool* activated)
 {
     BEGIN_HRESULT();
+
+    if (!ctx) return E_INVALIDARG;
 
     CHECK_HR(ctx->m_InputContext->getActivated(activated));
 
@@ -94,6 +116,8 @@ LIBTF_EXPORT HRESULT libtf_set_fullscreen(libtf_pInputContext ctx, const bool fu
 {
     BEGIN_HRESULT();
 
+    if (!ctx) return E_INVALIDARG;
+
     CHECK_HR(ctx->m_InputContext->m_FullScreenUIElementHandler->setFullScreen(fullscreen));
 
     END_HRESULT();
@@ -107,6 +131,8 @@ LIBTF_EXPORT HRESULT libtf_set_fullscreen(libtf_pInputContext ctx, const bool fu
 LIBTF_EXPORT HRESULT libtf_get_fullscreen(const libtf_pInputContext ctx, bool* fullscreen)
 {
     BEGIN_HRESULT();
+
+    if (!ctx) return E_INVALIDARG;
 
     CHECK_HR(ctx->m_InputContext->m_FullScreenUIElementHandler->getFullScreen(fullscreen));
 
@@ -124,52 +150,55 @@ LIBTF_EXPORT libtf_CandidateListCallback libtf_candidate_list_set_callback(libtf
                                                                            libtf_CandidateListCallback callback,
                                                                            void*                       userData)
 {
-    static libtf_CandidateListCallback PrevCallback = NULL;
-    static void*                       PrevUserData = NULL;
+    if (!ctx) return NULL;
 
-    auto prev    = PrevCallback;
-    PrevCallback = callback;
+    auto prev                        = ctx->m_PrevCandidateListCallback;
+    ctx->m_PrevCandidateListCallback = callback;
 
-    ctx->m_InputContext->m_FullScreenUIElementHandler->m_CandidateListHandler->setCallback(
-        [callback, userData](auto&& state, auto&& ctx) {
-            // NULL context only state update
-            if (!ctx) {
-                callback(state, NULL, userData);
-                return;
-            }
+    if (callback)
+        ctx->m_InputContext->m_FullScreenUIElementHandler->m_CandidateListHandler->setCallback(
+            [callback, userData](auto&& state, auto&& ctx) {
+                // NULL context only state update
+                if (!ctx) {
+                    callback(state, NULL, userData);
+                    return;
+                }
 
-            auto strTotalLen = std::accumulate(ctx->m_Candidates.begin(),
-                                               ctx->m_Candidates.end(),
-                                               ctx->m_PageSize,// Number of \0
-                                               [](auto&& sum, auto&& it) { return sum + (uint32_t)it.length(); });
-            auto libtf_ctx   = std::unique_ptr<libtf_CandidateListContext_t, decltype(&::free)>(
-                (libtf_pCandidateListContext)malloc(sizeof(libtf_CandidateListContext_t) +
-                                                    sizeof(wchar_t*) * ctx->m_PageSize + sizeof(wchar_t) * strTotalLen),
-                free);
+                auto strTotalLen = std::accumulate(ctx->m_Candidates.begin(),
+                                                   ctx->m_Candidates.end(),
+                                                   ctx->m_PageSize,// Number of \0
+                                                   [](auto&& sum, auto&& it) { return sum + (uint32_t)it.length(); });
+                auto libtf_ctx   = std::unique_ptr<libtf_CandidateListContext_t, decltype(&::free)>(
+                    (libtf_pCandidateListContext)malloc(sizeof(libtf_CandidateListContext_t) +
+                                                        sizeof(wchar_t*) * ctx->m_PageSize +
+                                                        sizeof(wchar_t) * strTotalLen),
+                    free);
 
-            // Out of memory
-            if (!libtf_ctx) return;
+                // Out of memory
+                if (!libtf_ctx) return;
 
-            libtf_ctx->m_Selection = ctx->m_Selection;
-            libtf_ctx->m_PageSize  = ctx->m_PageSize;
+                libtf_ctx->m_Selection = ctx->m_Selection;
+                libtf_ctx->m_PageSize  = ctx->m_PageSize;
 
-            // Append strings at the end of the struct
-            int  i           = 0;
-            auto pCandidates = (wchar_t*)&libtf_ctx->m_Candidates[ctx->m_PageSize];
-            for (auto&& it : ctx->m_Candidates) {
-                // Point to string start
-                libtf_ctx->m_Candidates[i++] = pCandidates;
+                // Append strings at the end of the struct
+                int  i           = 0;
+                auto pCandidates = (wchar_t*)&libtf_ctx->m_Candidates[ctx->m_PageSize];
+                for (auto&& it : ctx->m_Candidates) {
+                    // Point to string start
+                    libtf_ctx->m_Candidates[i++] = pCandidates;
 
-                auto size = it.length() + 1;
-                // Copy string data
-                memcpy(pCandidates, it.c_str(), size * sizeof(wchar_t));
-                pCandidates += size;
-            }
+                    auto size = it.length() + 1;
+                    // Copy string data
+                    memcpy(pCandidates, it.c_str(), size * sizeof(wchar_t));
+                    pCandidates += size;
+                }
 
-            callback(state, libtf_ctx.get(), userData);
-        });
+                callback(state, libtf_ctx.get(), userData);
+            });
+    else
+        ctx->m_InputContext->m_FullScreenUIElementHandler->m_CandidateListHandler->setCallback(NULL);
 
-    std::swap(PrevUserData, userData);
+    std::swap(ctx->m_PrevCandidateListUserData, userData);
 
     return prev;
 }
@@ -184,6 +213,8 @@ LIBTF_EXPORT HRESULT libtf_candidate_list_set_sel(const libtf_pInputContext ctx,
 {
     BEGIN_HRESULT();
 
+    if (!ctx) return E_INVALIDARG;
+
     auto candCtx = ctx->m_InputContext->m_FullScreenUIElementHandler->m_CandidateListHandler->m_Context;
     if (candCtx) CHECK_HR(candCtx->select(index));
 
@@ -197,6 +228,8 @@ LIBTF_EXPORT HRESULT libtf_candidate_list_set_sel(const libtf_pInputContext ctx,
 LIBTF_EXPORT HRESULT libtf_candidate_list_finalize(const libtf_pInputContext ctx)
 {
     BEGIN_HRESULT();
+
+    if (!ctx) return E_INVALIDARG;
 
     auto candCtx = ctx->m_InputContext->m_FullScreenUIElementHandler->m_CandidateListHandler->m_Context;
     if (candCtx) CHECK_HR(candCtx->finalize());
@@ -214,6 +247,8 @@ LIBTF_EXPORT HRESULT libtf_composition_terminate(const libtf_pInputContext ctx)
 {
     BEGIN_HRESULT();
 
+    if (!ctx) return E_INVALIDARG;
+
     CHECK_HR(ctx->m_InputContext->m_CompositionHandler->terminate());
 
     END_HRESULT();
@@ -230,41 +265,44 @@ LIBTF_EXPORT libtf_PreEditCallback libtf_preedit_set_callback(libtf_pInputContex
                                                               libtf_PreEditCallback callback,
                                                               void*                 userData)
 {
-    static libtf_PreEditCallback PrevCallback = NULL;
-    static void*                 PrevUserData = NULL;
+    if (!ctx) return NULL;
 
-    auto prev    = PrevCallback;
-    PrevCallback = callback;
+    auto prev                  = ctx->m_PrevPreEditCallback;
+    ctx->m_PrevPreEditCallback = callback;
 
-    ctx->m_InputContext->m_CompositionHandler->m_PreEditHandler->libtf::PreEditContextCallback::setCallback(
-        [callback, userData](auto&& state, auto&& ctx) {
-            // NULL context only state update
-            if (!ctx) {
-                callback(state, NULL, userData);
-                return;
-            }
+    if (callback)
+        ctx->m_InputContext->m_CompositionHandler->m_PreEditHandler->libtf::PreEditContextCallback::setCallback(
+            [callback, userData](auto&& state, auto&& ctx) {
+                // NULL context only state update
+                if (!ctx) {
+                    callback(state, NULL, userData);
+                    return;
+                }
 
-            auto strTotalLen = ctx->m_Content.length() + 1;
-            auto libtf_ctx   = std::unique_ptr<libtf_PreEditContext_t, decltype(&::free)>(
-                (libtf_pPreEditContext)malloc(sizeof(libtf_PreEditContext_t) + sizeof(wchar_t) * strTotalLen), free);
+                auto strTotalLen = ctx->m_Content.length() + 1;
+                auto libtf_ctx   = std::unique_ptr<libtf_PreEditContext_t, decltype(&::free)>(
+                    (libtf_pPreEditContext)malloc(sizeof(libtf_PreEditContext_t) + sizeof(wchar_t) * strTotalLen),
+                    free);
 
-            // Out of memory
-            if (!libtf_ctx) return;
+                // Out of memory
+                if (!libtf_ctx) return;
 
-            libtf_ctx->m_SelStart = ctx->m_SelStart;
-            libtf_ctx->m_SelEnd   = ctx->m_SelEnd;
+                libtf_ctx->m_SelStart = ctx->m_SelStart;
+                libtf_ctx->m_SelEnd   = ctx->m_SelEnd;
 
-            // Append string at the end of the struct
-            auto pContent = libtf_ctx->m_Content;
+                // Append string at the end of the struct
+                auto pContent = libtf_ctx->m_Content;
 
-            auto size = (ctx->m_Content.length() + 1) * sizeof(wchar_t);
-            // Copy string data
-            memcpy(pContent, ctx->m_Content.c_str(), size);
+                auto size = (ctx->m_Content.length() + 1) * sizeof(wchar_t);
+                // Copy string data
+                memcpy(pContent, ctx->m_Content.c_str(), size);
 
-            callback(state, libtf_ctx.get(), userData);
-        });
+                callback(state, libtf_ctx.get(), userData);
+            });
+    else
+        ctx->m_InputContext->m_CompositionHandler->m_PreEditHandler->libtf::PreEditContextCallback::setCallback(NULL);
 
-    std::swap(PrevUserData, userData);
+    std::swap(ctx->m_PrevPreEditUserData, userData);
 
     return prev;
 }
@@ -281,16 +319,18 @@ LIBTF_EXPORT libtf_PreEditRectCallback libtf_preedit_rect_set_callback(libtf_pIn
                                                                        libtf_PreEditRectCallback callback,
                                                                        void*                     userData)
 {
-    static libtf_PreEditRectCallback PrevCallback = NULL;
-    static void*                     PrevUserData = NULL;
+    if (!ctx) return NULL;
 
-    auto prev    = PrevCallback;
-    PrevCallback = callback;
+    auto prev                      = ctx->m_PrevPreEditRectCallback;
+    ctx->m_PrevPreEditRectCallback = callback;
 
-    ctx->m_InputContext->m_CompositionHandler->m_PreEditHandler->libtf::PreEditRectCallback::setCallback(
-        [callback, userData](auto&& it) { callback(it, userData); });
+    if (callback)
+        ctx->m_InputContext->m_CompositionHandler->m_PreEditHandler->libtf::PreEditRectCallback::setCallback(
+            [callback, userData](auto&& it) { callback(it, userData); });
+    else
+        ctx->m_InputContext->m_CompositionHandler->m_PreEditHandler->libtf::PreEditRectCallback::setCallback(NULL);
 
-    std::swap(PrevUserData, userData);
+    std::swap(ctx->m_PrevPreEditRectUserData, userData);
 
     return prev;
 }
@@ -303,16 +343,18 @@ LIBTF_EXPORT libtf_CommitCallback libtf_commit_set_callback(libtf_pInputContext 
                                                             libtf_CommitCallback callback,
                                                             void*                userData)
 {
-    static libtf_CommitCallback PrevCallback = NULL;
-    static void*                PrevUserData = NULL;
+    if (!ctx) return NULL;
 
-    auto prev    = PrevCallback;
-    PrevCallback = callback;
+    auto prev                 = ctx->m_PrevCommitCallback;
+    ctx->m_PrevCommitCallback = callback;
 
-    ctx->m_InputContext->m_CompositionHandler->m_CommitHandler->setCallback(
-        [callback, userData](auto&& it) { callback(it.c_str(), userData); });
+    if (callback)
+        ctx->m_InputContext->m_CompositionHandler->m_CommitHandler->setCallback(
+            [callback, userData](auto&& it) { callback(it.c_str(), userData); });
+    else
+        ctx->m_InputContext->m_CompositionHandler->m_CommitHandler->setCallback(NULL);
 
-    std::swap(PrevUserData, userData);
+    std::swap(ctx->m_PrevCommitUserData, userData);
 
     return prev;
 }
@@ -366,19 +408,21 @@ LIBTF_EXPORT libtf_InputProcessorCallback libtf_inputprocessor_set_callback(libt
                                                                             libtf_InputProcessorCallback callback,
                                                                             void*                        userData)
 {
-    static libtf_InputProcessorCallback PrevCallback = NULL;
-    static void*                        PrevUserData = NULL;
+    if (!ctx) return NULL;
 
-    auto prev    = PrevCallback;
-    PrevCallback = callback;
+    auto prev                         = ctx->m_PrevInputProcessorCallback;
+    ctx->m_PrevInputProcessorCallback = callback;
 
-    ctx->m_InputContext->m_InputProcessorHandler->setCallback([callback, userData](auto&& state, auto&& ctx) {
-        auto libtf_ctx =
-            std::unique_ptr<libtf_InputProcessorContext_t, decltype(&::free)>(inputprocessor_get_ctx(*ctx), free);
-        if (libtf_ctx) callback(state, libtf_ctx.get(), userData);
-    });
+    if (callback)
+        ctx->m_InputContext->m_InputProcessorHandler->setCallback([callback, userData](auto&& state, auto&& ctx) {
+            auto libtf_ctx =
+                std::unique_ptr<libtf_InputProcessorContext_t, decltype(&::free)>(inputprocessor_get_ctx(*ctx), free);
+            if (libtf_ctx) callback(state, libtf_ctx.get(), userData);
+        });
+    else
+        ctx->m_InputContext->m_InputProcessorHandler->setCallback(NULL);
 
-    std::swap(PrevUserData, userData);
+    std::swap(ctx->m_PrevInputProcessorUserData, userData);
 
     return prev;
 }
@@ -390,6 +434,8 @@ LIBTF_EXPORT libtf_InputProcessorCallback libtf_inputprocessor_set_callback(libt
  */
 LIBTF_EXPORT libtf_pInputProcessorContext libtf_inputprocessor_get_ctx(const libtf_pInputContext ctx)
 {
+    if (!ctx) return NULL;
+
     auto processorCtx = ctx->m_InputContext->m_InputProcessorHandler->m_Context;
     return inputprocessor_get_ctx(*processorCtx);
 }
@@ -412,6 +458,8 @@ LIBTF_EXPORT void libtf_inputprocessor_free_ctx(libtf_pInputProcessorContext* ct
 LIBTF_EXPORT HRESULT libtf_inputprocessor_apply_inputmode(libtf_pInputContext ctx, const wchar_t* newMode)
 {
     BEGIN_HRESULT();
+
+    if (!ctx) return E_INVALIDARG;
 
     CHECK_HR(ctx->m_InputContext->m_InputProcessorHandler->applyInputMode(newMode));
 
@@ -462,6 +510,8 @@ LIBTF_EXPORT void libtf_free_inputprocessors(libtf_pInputProcessors* processors)
  */
 LIBTF_EXPORT libtf_pInputProcessorProfile libtf_inputprocessor_get_profile(const libtf_HInputProcessor hInputProcessor)
 {
+    if (!hInputProcessor) return NULL;
+
     auto processor   = ((const libtf::InputProcessor*)hInputProcessor);
     auto strTotalLen = processor->m_InputProcessorName.length();
     strTotalLen += processor->m_Locale.length();
@@ -519,6 +569,8 @@ LIBTF_EXPORT void libtf_inputprocessor_free_profile(libtf_pInputProcessorProfile
 LIBTF_EXPORT HRESULT libtf_inputprocessor_set_activated(const libtf_HInputProcessor hInputProcessor)
 {
     BEGIN_HRESULT();
+
+    if (!hInputProcessor) return E_INVALIDARG;
 
     CHECK_HR(((libtf::InputProcessor*)hInputProcessor)->setActivated());
